What: refactor app struct
Why:

  * prepair for rpc

How:

  * gogo new myapp
