+++ cmd/commands/application.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path"
-	"strings"
 
 	"github.com/golib/cli"
 )
@@ -16,7 +15,9 @@ var (
 		{"app", "controllers"},
 		{"app", "middlewares"},
 		{"app", "models"},
+		{"app", "protos"},
 		{"config"},
+		{"gogo", "service"},
 		{"lib"},
 		{"log"},
 		{"tmp", "cache"},
@@ -45,16 +46,10 @@ func (_ *_Application) Flags() []cli.Flag {
 			Usage:  "specify application package import path, default to github.com",
 			EnvVar: "GOGO_NAMESPACE",
 		},
-		cli.StringFlag{
-			Name:   "project",
-			Value:  "gogo-project",
-			Usage:  "specify application project name, default to gogo-project",
-			EnvVar: "GOGO_PROJECT",
-		},
 		cli.BoolFlag{
-			Name:   "skip-install",
-			Usage:  "skip `go get` dependences installation",
-			EnvVar: "GOGO_SKIP_INSTALL",
+			Name:   "go-install",
+			Usage:  "run `go mod tidy` for dependences resolving",
+			EnvVar: "GOGO_GO_INSTALL",
 		},
 	}
 }
@@ -63,151 +58,92 @@ func (_ *_Application) Action() cli.ActionFunc {
 	return func(ctx *cli.Context) error {
 		root, err := os.Getwd()
 		if err != nil {
-			stderr.Error(err.Error())
+			log.Error(err.Error())
 
 			return err
 		}
 
-		appName := path.Clean(ctx.Args().First())
-		if appName != "" {
-			root += "/" + strings.TrimPrefix(appName, "/")
+		name := path.Clean(ctx.Args().First())
+		if name != "" {
+			root = path.Join(root, name)
 		}
 
 		// is app root is empty?
 		files, err := ioutil.ReadDir(root)
 		if err != nil && !os.IsNotExist(err) {
-			stderr.Error(err.Error())
+			log.Error(err.Error())
 
 			return err
 		}
-
 		if len(files) > 0 {
-			stderr.Warn(ErrNoneEmptyDirectory.Error())
+			log.Warn(ErrNoneEmptyDirectory.Error())
 
 			return ErrNoneEmptyDirectory
 		}
 
-		// create app root dir
-		appRoot := path.Join(root, "gogo")
-
-		err = os.MkdirAll(appRoot, os.ModePerm)
-		if err != nil {
-			stderr.Error(err.Error())
-
-			return err
-		}
-
 		// generate app struct
 		for _, dir := range appDirs {
-			absPath := path.Join(appRoot, path.Join(dir...))
+			dirname := path.Join(root, path.Join(dir...))
 
-			err := os.MkdirAll(absPath, os.ModePerm)
+			err := os.MkdirAll(dirname, os.ModePerm)
 			if err != nil {
-				stderr.Error(err.Error())
+				log.Error(err.Error())
 
 				return err
 			}
 
-			err = ioutil.WriteFile(path.Join(absPath, ".keep"), []byte(""), os.ModePerm)
+			err = ioutil.WriteFile(path.Join(dirname, ".keep"), []byte(""), os.ModePerm)
 			if err != nil {
-				stderr.Error(err.Error())
+				log.Error(err.Error())
 
 				return err
 			}
 		}
 
-		appNamespace := ctx.String("namespace")
+		namespace := ctx.String("namespace")
 
 		// generate .gitignore
-		Application.genGitIgnore(path.Join(root, ".gitignore"), appName, appNamespace)
+		Application.genGitIgnore(path.Join(root, ".gitignore"), name, namespace)
 
-		// generate env.sh
-		Application.genEnvFile(path.Join(root, "env.sh"), appName, appNamespace)
+		// // generate env.sh
+		// Application.genEnvfile(path.Join(root, "env.sh"), name, namespace)
 
 		// generate readme.md
-		Application.genReadme(path.Join(root, "README.md"), appName, appNamespace)
-
-		// generate go.mod
-		Application.genModfile(path.Join(appRoot, "go.mod"), appName, appNamespace)
+		Application.genReadme(path.Join(root, "README.md"), name, namespace)
 
 		// generate Makefile
-		Application.genMakefile(path.Join(appRoot, "Makefile"), appName, appNamespace)
+		Application.genMakefile(path.Join(root, "Makefile"), name, namespace)
+
+		// generate go.mod
+		Application.genModfile(path.Join(root, "go.mod"), name, namespace)
 
 		// generate default controller dependences
-		Application.genControllers(path.Join(appRoot, "app", "controllers"), appName, appNamespace)
+		Application.genControllers(path.Join(root, "app", "controllers"), name, namespace)
 
 		// generate default middlewares
-		Application.genMiddlewares(path.Join(appRoot, "app", "middlewares"), appName, appNamespace)
+		Application.genMiddlewares(path.Join(root, "app", "middlewares"), name, namespace)
 
 		// generate default models
-		Application.genModels(path.Join(appRoot, "app", "models"), appName, appNamespace)
+		Application.genModels(path.Join(root, "app", "models"), name, namespace)
 
 		// generate default application.json
-		Application.genConfigFile(path.Join(appRoot, "config", "application.json"), appName, appNamespace)
+		Application.genConfigFile(path.Join(root, "config", "application.json"), name, namespace)
 
 		// generate main.go
-		Application.genMainFile(path.Join(appRoot, "main.go"), appName, appNamespace)
+		Application.genMainFile(path.Join(root, "app", "main.go"), name, namespace)
 
-		// // auto install dependences
-		// if !ctx.Bool("skip-install") {
-		// 	Application.getDependences()
-		// }
+		// auto install dependences
+		if ctx.Bool("go-install") {
+			Application.runGoInstall(root)
+		}
 		return nil
-	}
-}
-
-func (_ *_Application) genEnvFile(file, app, namespace string) {
-	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		stderr.Error(err.Error())
-		return
-	}
-
-	err = box.Lookup("env").Execute(fd, templateData{
-		Namespace:   namespace,
-		Application: app,
-	})
-	if err != nil {
-		stderr.Error(err.Error())
-	}
-}
-
-func (_ *_Application) genModfile(file, app, namespace string) {
-	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		stderr.Error(err.Error())
-		return
-	}
-
-	err = box.Lookup("go.mod").Execute(fd, templateData{
-		Namespace:   namespace,
-		Application: app,
-	})
-	if err != nil {
-		stderr.Error(err.Error())
-	}
-}
-
-func (_ *_Application) genMakefile(file, app, namespace string) {
-	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		stderr.Error(err.Error())
-		return
-	}
-
-	err = box.Lookup("makefile").Execute(fd, templateData{
-		Namespace:   namespace,
-		Application: app,
-	})
-	if err != nil {
-		stderr.Error(err.Error())
 	}
 }
 
 func (_ *_Application) genGitIgnore(file, app, namespace string) {
 	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
 
@@ -216,14 +152,14 @@ func (_ *_Application) genGitIgnore(file, app, namespace string) {
 		Application: app,
 	})
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 	}
 }
 
 func (_ *_Application) genReadme(file, app, namespace string) {
 	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
 
@@ -232,7 +168,55 @@ func (_ *_Application) genReadme(file, app, namespace string) {
 		Application: app,
 	})
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
+	}
+}
+
+func (_ *_Application) genMakefile(file, app, namespace string) {
+	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		log.Error(err.Error())
+		return
+	}
+
+	err = box.Lookup("makefile").Execute(fd, templateData{
+		Namespace:   namespace,
+		Application: app,
+	})
+	if err != nil {
+		log.Error(err.Error())
+	}
+}
+
+func (_ *_Application) genEnvfile(file, app, namespace string) {
+	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		log.Error(err.Error())
+		return
+	}
+
+	err = box.Lookup("env.sh").Execute(fd, templateData{
+		Namespace:   namespace,
+		Application: app,
+	})
+	if err != nil {
+		log.Error(err.Error())
+	}
+}
+
+func (_ *_Application) genModfile(file, app, namespace string) {
+	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		log.Error(err.Error())
+		return
+	}
+
+	err = box.Lookup("go.mod").Execute(fd, templateData{
+		Namespace:   namespace,
+		Application: app,
+	})
+	if err != nil {
+		log.Error(err.Error())
 	}
 }
 
@@ -245,73 +229,67 @@ func (_ *_Application) genControllers(root, app, namespace string) {
 	// application.go
 	fd, err := os.OpenFile(path.Join(root, "application.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("application").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("application.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 
 	// testing_test.go
 	fd, err = os.OpenFile(path.Join(root, "testing_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("application_testing").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("application_testing.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 
 	// application_config.go
 	fd, err = os.OpenFile(path.Join(root, "application_config.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("application_config").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("application_config.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 
 	// application_config_test.go
 	fd, err = os.OpenFile(path.Join(root, "application_config_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("application_config_test").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("application_config_test.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 
 	// getting_start.go
 	fd, err = os.OpenFile(path.Join(root, "getting_start.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("getting_start").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("getting_start.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 
 	// getting_start_test.go
 	fd, err = os.OpenFile(path.Join(root, "getting_start_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("getting_start_test").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("getting_start_test.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
+		return
 	}
 }
 
@@ -324,37 +302,31 @@ func (_ *_Application) genMiddlewares(root, app, namespace string) {
 	// testing_test.go
 	fd, err := os.OpenFile(path.Join(root, "testing_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("middleware_testing").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("middleware_testing.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
 	}
 
 	// recovery.go
 	fd, err = os.OpenFile(path.Join(root, "recovery.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("middleware_recovery").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("middleware_recovery.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
 	}
 
 	// recovery_test.go
 	fd, err = os.OpenFile(path.Join(root, "recovery_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("middleware_recovery_test").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("middleware_recovery_test.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
 	}
 }
 
@@ -367,56 +339,56 @@ func (_ *_Application) genModels(root, app, namespace string) {
 	// model.go
 	fd, err := os.OpenFile(path.Join(root, "model.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("model").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("model.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
 	}
 
 	// model_test.go
 	fd, err = os.OpenFile(path.Join(root, "model_test.go"), os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
-
-	err = box.Lookup("model_test").Execute(fd, data)
-	if err != nil {
-		stderr.Error(err.Error())
+	if err := box.Lookup("model_test.go").Execute(fd, data); err != nil {
+		log.Error(err.Error())
 	}
 }
 
 func (_ *_Application) genConfigFile(file, app, namespace string) {
 	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
 
-	err = box.Lookup("application_config_json").Execute(fd, templateData{
+	err = box.Lookup("application_config.json").Execute(fd, templateData{
 		Namespace:   namespace,
 		Application: app,
 	})
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 	}
 }
 
 func (_ *_Application) genMainFile(file, app, namespace string) {
 	fd, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 		return
 	}
 
-	err = box.Lookup("main").Execute(fd, templateData{
+	err = box.Lookup("main.go").Execute(fd, templateData{
 		Namespace:   namespace,
 		Application: app,
 	})
 	if err != nil {
-		stderr.Error(err.Error())
+		log.Error(err.Error())
 	}
+}
+
+func (_ *_Application) runGoInstall(root string) {
+	// TODO: run go mod tidy auto
 }
