--- cmd/gogo/templates/doc.go
@@ -1,8 +1,0 @@
-// Copyright 2009 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-/*
-	Package templates provides templates for gogo application generation.
-*/
-package templates
