--- cmd/gogo/templates/help.txt
@@ -1,1 +1,0 @@
-Hello, gogo!
