+++ cmd/gogo/templates/mod.go
@@ -1,0 +1,6 @@
+package templates
+
+var (
+	modTemplate = `module {{.Namespace}}/{{.Application}}
+`
+)
