+++ gogo_test.go
@@ -15,7 +15,7 @@ var (
 	fakeApp = func(mode string) *AppServer {
 		root, _ := os.Getwd()
 
-		return New(mode, path.Join(root, "skeleton", "gogo"))
+		return New(mode, path.Join(root, "skeleton"))
 	}
 )
