+++ skeleton/Makefile
@@ -1,7 +1,9 @@
-all: gobuild gotest
+all: gobuild gotest gopackage
+
+dev: gobuild godev
 
 godev:
-	cd gogo && go run main.go
+	go run app/main.go
 
 gobuild: goclean goinstall
 
@@ -12,11 +14,13 @@ goclean:
 
 goinstall:
 	go get github.com/dolab/httptesting
+	go get github.com/dolab/httpmitm
 	go get github.com/dolab/logger
 	go get github.com/golib/assert
 
 goreinstall:
 	go get -u github.com/dolab/httptesting
+	go get -u github.com/dolab/httpmitm
 	go get -u github.com/dolab/logger
 	go get -u github.com/golib/assert
 
@@ -26,6 +30,6 @@ gotest:
 	go test github.com/skeleton/app/models
 
 gopackage:
-	mkdir -p bin && go build -a -o bin/skeleton src/github.com/skeleton/gogo/main.go
+	mkdir -p bin && go build -a -o bin/skeleton app/main.go
 
 travis: gobuild gotest
