+++ skeleton/app/controllers/getting_start_test.go
@@ -6,7 +6,7 @@ import (
 
 func Test_GettingStart_Hello(t *testing.T) {
 	request := gogotesting.New(t)
-	request.Get("/@getting_start/hello")
+	request.Get("/@gretting/hello")
 
 	request.AssertOK()
 	request.AssertContains(Config.GettingStart.Greeting)
