+++ skeleton/app/controllers/testing_test.go
@@ -1,0 +1,31 @@
+package controllers
+
+import (
+	"os"
+	"path"
+	"testing"
+
+	"github.com/dolab/httptesting"
+)
+
+var (
+	gogotesting *httptesting.Client
+)
+
+func TestMain(m *testing.M) {
+	var (
+		runMode = "test"
+		srcPath = path.Clean("../../")
+	)
+
+	app := gogo.New(runMode, srcPath)
+	app.NewResources(New())
+
+	gogotesting = httptesting.NewServer(app, false)
+
+	code := m.Run()
+
+	gogotesting.Close()
+
+	os.Exit(code)
+}
