+++ skeleton/app/models/model.go
