+++ skeleton/config/application.healthz.json
