+++ skeleton/config/application.json
