+++ skeleton/go.mod
@@ -1,0 +1,8 @@
+module github.com/myapp
+
+require (
+	github.com/dolab/gogo v0.0.0-20190122083835-0f0d9fc2581f
+	github.com/dolab/httpmitm v0.0.0-20181102123249-38079d87cfa7 // indirect
+	github.com/dolab/httptesting v0.0.0-20181015062713-fea093b8a92e
+	github.com/golib/assert v0.0.0-20170825111607-0306abba9bd3
+)
