--- skeleton/gogo/config/application.development.json
@@ -1,30 +1,0 @@
-{
-    "name": "for development",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "ssl_cert": "/path/to/ssl/cert",
-                "ssl_key": "/path/to/ssl/key",
-                "request_timeout": 30,
-                "response_timeout": 30,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            },
-            "domain": "https://example.com",
-            "getting_start": {
-                "greeting": "Hello, gogo!"
-            }
-        }
-    }
-}
